+++ app/src/main/AndroidManifest.xml
@@ -11,7 +11,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".com.swmaestro.roundup.home.HomeFeedActivity"
+            android:name=".home.HomeFeedActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
@@ -21,13 +21,13 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".com.swmaestro.roundup.navigation.AddGroupActivity">
+            android:name=".navigation.AddGroupActivity">
         </activity>
         <activity
-            android:name=".com.swmaestro.roundup.club.ClubActivity"
+            android:name=".club.ClubActivity"
             android:theme="@style/AppTheme.NoActionBar">
         </activity>
-        <activity android:name=".com.swmaestro.roundup.club.AddPostActivity"/>
+        <activity android:name=".club.AddPostActivity"/>
     </application>
 
 </manifest>
