+++ app/src/main/java/com/swmaestro/roundup/club/PostEventFragment.java
@@ -1,4 +1,4 @@
-package com.swmaestro.roundup.com.swmaestro.roundup.club;
+package com.swmaestro.roundup.club;
 
 import android.app.Fragment;
 import android.os.Bundle;
@@ -11,13 +11,12 @@ import com.swmaestro.roundup.R;
 /**
  * Created by lk on 16. 5. 19..
  */
-public class PostSessionFragment extends Fragment {
+public class PostEventFragment extends Fragment {
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
 
-        return inflater.inflate(R.layout.fragment_club_post_session, container, false);
+        return inflater.inflate(R.layout.fragment_club_post_event, container, false);
     }
-
 }
