+++ app/src/main/java/com/swmaestro/roundup/home/HomeFeedActivity.java
@@ -1,4 +1,4 @@
-package com.swmaestro.roundup.com.swmaestro.roundup.home;
+package com.swmaestro.roundup.home;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -16,9 +16,9 @@ import android.support.v7.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuItem;
 
-import com.swmaestro.roundup.com.swmaestro.roundup.club.ClubActivity;
-import com.swmaestro.roundup.com.swmaestro.roundup.navigation.AddGroupActivity;
 import com.swmaestro.roundup.R;
+import com.swmaestro.roundup.club.ClubActivity;
+import com.swmaestro.roundup.navigation.AddGroupActivity;
 
 public class HomeFeedActivity extends AppCompatActivity
         implements NavigationView.OnNavigationItemSelectedListener {
